+++ tests/Unit/Framework/Attribute/Providers/ProviderDefaultEmpty.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Unit\Framework\Attribute\Providers;
+
+use Gacela\Framework\AbstractProvider;
+
+final class ProviderDefaultEmpty extends AbstractProvider
+{
+}